+++ tools/curleo/main.go
@@ -72,12 +72,15 @@ func main() {
 	}
 	metadata := curl.parseMetadata()
 	spender := curl.getSpender(*user)
+
+	chainID := flag.String("chain-id", "arkeo", "chain ID to use for signing")
+
 	contract := curl.getActiveContract(metadata.Configuration.ProviderPubKey.String(), service, spender)
 	if contract.Height == 0 {
-		println(fmt.Sprintf("no active contract found for provider:%s cbhain:%s - will attempt free tier", metadata.Configuration.ProviderPubKey.String(), service))
+		println(fmt.Sprintf("no active contract found for provider:%s chain:%s - will attempt free tier", metadata.Configuration.ProviderPubKey.String(), service))
 	} else {
 		claim := curl.getClaim(contract.Id)
-		auth := curl.sign(*user, contract.Id, claim.Nonce+1)
+		auth := curl.sign(*user, contract.Id, claim.Nonce+1, *chainID)
 		values.Add(sentinel.QueryArkAuth, auth)
 	}
 	u.RawQuery = values.Encode()
@@ -189,7 +192,7 @@ func (c Curl) parseMetadata() sentinel.Metadata {
 	return meta
 }
 
-func (c Curl) sign(user string, contractId uint64, nonce int64) string {
+func (c Curl) sign(user string, contractId uint64, nonce int64, chainId string) string {
 	interfaceRegistry := codectypes.NewInterfaceRegistry()
 	std.RegisterInterfaces(interfaceRegistry)
 	ModuleBasics.RegisterInterfaces(interfaceRegistry)
@@ -203,7 +206,7 @@ func (c Curl) sign(user string, contractId uint64, nonce int64) string {
 	if err != nil {
 		log.Fatal(err)
 	}
-	msg := sentinel.GenerateMessageToSign(contractId, nonce)
+	msg := sentinel.GenerateMessageToSign(contractId, nonce, chainId)
 
 	println("invoking Sign...")
 	signature, pk, err := kb.Sign(user, []byte(msg), signing.SignMode_SIGN_MODE_DIRECT)