+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Contains all the PRs that improved the code without changing the behaviors.
 - Fixed bug allowing double claims with Thorchain claims.
 - Fixed non deterministic map iteration to sorted iteration 
 - Fixed fee stuck in arkeo module by moving fees to arkeo-reserve
+- Fixed signature replay on different network
 
 ---